+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.7.8
 - Fixed disable protection modal positioning
+- moved json files location to AppData Roaming WinOtp folder
 
 # v0.7.7
 - Fixed regression with next code preview, showing the same code after the first preview.